+++ Include/Acidanthera/Library/OcDriverConnectionLib.h
@@ -39,6 +39,14 @@ OcUnblockUnmountedPartitions (
   );
 
 /**
+  Disconnects all graphics drivers attached to PCI I/O protcols.
+**/
+VOID
+OcDisconnectGraphicsDrivers (
+  VOID
+  );
+
+/**
   Disconnect effectively all drivers attached at handle.
 
   @param[in] Controller    Handle to disconnect the drivers from.